requirements.txt
# Flask Framework
Flask[async]>=2.3.3,<3.0.0
Flask>=2.3.3,<4.0.0
werkzeug<4

# Flask Extensions
Flask-Login>=0.6
Flask-WTF>=1.1
Flask-SQLAlchemy>=3.0
Flask-Bcrypt==1.0
Flask-Migrate>=3.1

python-dotenv==1.0.0
email_validator
# psycopg2-binary==2.9.6

pdfminer.six[image]
pandas